Move rover gen into repo.
