--- apps/rover/src/casadi/bezier6.h
@@ -1,49 +1,0 @@
-/* This file was automatically generated by CasADi.
-   The CasADi copyright holders make no ownership claim of its contents. */
-#ifdef __cplusplus
-extern "C" {
-#endif
-
-#ifndef casadi_real
-#define casadi_real double
-#endif
-
-#ifndef casadi_int
-#define casadi_int long long int
-#endif
-
-int bezier6_solve(const casadi_real** arg, casadi_real** res, casadi_int* iw, casadi_real* w, int mem);
-int bezier6_solve_alloc_mem(void);
-int bezier6_solve_init_mem(int mem);
-void bezier6_solve_free_mem(int mem);
-int bezier6_solve_checkout(void);
-void bezier6_solve_release(int mem);
-void bezier6_solve_incref(void);
-void bezier6_solve_decref(void);
-casadi_int bezier6_solve_n_in(void);
-casadi_int bezier6_solve_n_out(void);
-casadi_real bezier6_solve_default_in(casadi_int i);
-const char* bezier6_solve_name_in(casadi_int i);
-const char* bezier6_solve_name_out(casadi_int i);
-const casadi_int* bezier6_solve_sparsity_in(casadi_int i);
-const casadi_int* bezier6_solve_sparsity_out(casadi_int i);
-int bezier6_solve_work(casadi_int *sz_arg, casadi_int* sz_res, casadi_int *sz_iw, casadi_int *sz_w);
-int bezier6_traj(const casadi_real** arg, casadi_real** res, casadi_int* iw, casadi_real* w, int mem);
-int bezier6_traj_alloc_mem(void);
-int bezier6_traj_init_mem(int mem);
-void bezier6_traj_free_mem(int mem);
-int bezier6_traj_checkout(void);
-void bezier6_traj_release(int mem);
-void bezier6_traj_incref(void);
-void bezier6_traj_decref(void);
-casadi_int bezier6_traj_n_in(void);
-casadi_int bezier6_traj_n_out(void);
-casadi_real bezier6_traj_default_in(casadi_int i);
-const char* bezier6_traj_name_in(casadi_int i);
-const char* bezier6_traj_name_out(casadi_int i);
-const casadi_int* bezier6_traj_sparsity_in(casadi_int i);
-const casadi_int* bezier6_traj_sparsity_out(casadi_int i);
-int bezier6_traj_work(casadi_int *sz_arg, casadi_int* sz_res, casadi_int *sz_iw, casadi_int *sz_w);
-#ifdef __cplusplus
-} /* extern "C" */
-#endif
