+++ apps/rover/src/casadi/rover.h
@@ -12,22 +12,134 @@ extern "C" {
 #define casadi_int long long int
 #endif
 
-int rover(const casadi_real** arg, casadi_real** res, casadi_int* iw, casadi_real* w, int mem);
-int rover_alloc_mem(void);
-int rover_init_mem(int mem);
-void rover_free_mem(int mem);
-int rover_checkout(void);
-void rover_release(int mem);
-void rover_incref(void);
-void rover_decref(void);
-casadi_int rover_n_in(void);
-casadi_int rover_n_out(void);
-casadi_real rover_default_in(casadi_int i);
-const char* rover_name_in(casadi_int i);
-const char* rover_name_out(casadi_int i);
-const casadi_int* rover_sparsity_in(casadi_int i);
-const casadi_int* rover_sparsity_out(casadi_int i);
-int rover_work(casadi_int *sz_arg, casadi_int* sz_res, casadi_int *sz_iw, casadi_int *sz_w);
+int bezier6_solve(const casadi_real** arg, casadi_real** res, casadi_int* iw, casadi_real* w, int mem);
+int bezier6_solve_alloc_mem(void);
+int bezier6_solve_init_mem(int mem);
+void bezier6_solve_free_mem(int mem);
+int bezier6_solve_checkout(void);
+void bezier6_solve_release(int mem);
+void bezier6_solve_incref(void);
+void bezier6_solve_decref(void);
+casadi_int bezier6_solve_n_in(void);
+casadi_int bezier6_solve_n_out(void);
+casadi_real bezier6_solve_default_in(casadi_int i);
+const char* bezier6_solve_name_in(casadi_int i);
+const char* bezier6_solve_name_out(casadi_int i);
+const casadi_int* bezier6_solve_sparsity_in(casadi_int i);
+const casadi_int* bezier6_solve_sparsity_out(casadi_int i);
+int bezier6_solve_work(casadi_int *sz_arg, casadi_int* sz_res, casadi_int *sz_iw, casadi_int *sz_w);
+int bezier6_traj(const casadi_real** arg, casadi_real** res, casadi_int* iw, casadi_real* w, int mem);
+int bezier6_traj_alloc_mem(void);
+int bezier6_traj_init_mem(int mem);
+void bezier6_traj_free_mem(int mem);
+int bezier6_traj_checkout(void);
+void bezier6_traj_release(int mem);
+void bezier6_traj_incref(void);
+void bezier6_traj_decref(void);
+casadi_int bezier6_traj_n_in(void);
+casadi_int bezier6_traj_n_out(void);
+casadi_real bezier6_traj_default_in(casadi_int i);
+const char* bezier6_traj_name_in(casadi_int i);
+const char* bezier6_traj_name_out(casadi_int i);
+const casadi_int* bezier6_traj_sparsity_in(casadi_int i);
+const casadi_int* bezier6_traj_sparsity_out(casadi_int i);
+int bezier6_traj_work(casadi_int *sz_arg, casadi_int* sz_res, casadi_int *sz_iw, casadi_int *sz_w);
+int bezier6_rover(const casadi_real** arg, casadi_real** res, casadi_int* iw, casadi_real* w, int mem);
+int bezier6_rover_alloc_mem(void);
+int bezier6_rover_init_mem(int mem);
+void bezier6_rover_free_mem(int mem);
+int bezier6_rover_checkout(void);
+void bezier6_rover_release(int mem);
+void bezier6_rover_incref(void);
+void bezier6_rover_decref(void);
+casadi_int bezier6_rover_n_in(void);
+casadi_int bezier6_rover_n_out(void);
+casadi_real bezier6_rover_default_in(casadi_int i);
+const char* bezier6_rover_name_in(casadi_int i);
+const char* bezier6_rover_name_out(casadi_int i);
+const casadi_int* bezier6_rover_sparsity_in(casadi_int i);
+const casadi_int* bezier6_rover_sparsity_out(casadi_int i);
+int bezier6_rover_work(casadi_int *sz_arg, casadi_int* sz_res, casadi_int *sz_iw, casadi_int *sz_w);
+int ackermann_steering(const casadi_real** arg, casadi_real** res, casadi_int* iw, casadi_real* w, int mem);
+int ackermann_steering_alloc_mem(void);
+int ackermann_steering_init_mem(int mem);
+void ackermann_steering_free_mem(int mem);
+int ackermann_steering_checkout(void);
+void ackermann_steering_release(int mem);
+void ackermann_steering_incref(void);
+void ackermann_steering_decref(void);
+casadi_int ackermann_steering_n_in(void);
+casadi_int ackermann_steering_n_out(void);
+casadi_real ackermann_steering_default_in(casadi_int i);
+const char* ackermann_steering_name_in(casadi_int i);
+const char* ackermann_steering_name_out(casadi_int i);
+const casadi_int* ackermann_steering_sparsity_in(casadi_int i);
+const casadi_int* ackermann_steering_sparsity_out(casadi_int i);
+int ackermann_steering_work(casadi_int *sz_arg, casadi_int* sz_res, casadi_int *sz_iw, casadi_int *sz_w);
+int differential_steering(const casadi_real** arg, casadi_real** res, casadi_int* iw, casadi_real* w, int mem);
+int differential_steering_alloc_mem(void);
+int differential_steering_init_mem(int mem);
+void differential_steering_free_mem(int mem);
+int differential_steering_checkout(void);
+void differential_steering_release(int mem);
+void differential_steering_incref(void);
+void differential_steering_decref(void);
+casadi_int differential_steering_n_in(void);
+casadi_int differential_steering_n_out(void);
+casadi_real differential_steering_default_in(casadi_int i);
+const char* differential_steering_name_in(casadi_int i);
+const char* differential_steering_name_out(casadi_int i);
+const casadi_int* differential_steering_sparsity_in(casadi_int i);
+const casadi_int* differential_steering_sparsity_out(casadi_int i);
+int differential_steering_work(casadi_int *sz_arg, casadi_int* sz_res, casadi_int *sz_iw, casadi_int *sz_w);
+int se2_U(const casadi_real** arg, casadi_real** res, casadi_int* iw, casadi_real* w, int mem);
+int se2_U_alloc_mem(void);
+int se2_U_init_mem(int mem);
+void se2_U_free_mem(int mem);
+int se2_U_checkout(void);
+void se2_U_release(int mem);
+void se2_U_incref(void);
+void se2_U_decref(void);
+casadi_int se2_U_n_in(void);
+casadi_int se2_U_n_out(void);
+casadi_real se2_U_default_in(casadi_int i);
+const char* se2_U_name_in(casadi_int i);
+const char* se2_U_name_out(casadi_int i);
+const casadi_int* se2_U_sparsity_in(casadi_int i);
+const casadi_int* se2_U_sparsity_out(casadi_int i);
+int se2_U_work(casadi_int *sz_arg, casadi_int* sz_res, casadi_int *sz_iw, casadi_int *sz_w);
+int se2_U_inv(const casadi_real** arg, casadi_real** res, casadi_int* iw, casadi_real* w, int mem);
+int se2_U_inv_alloc_mem(void);
+int se2_U_inv_init_mem(int mem);
+void se2_U_inv_free_mem(int mem);
+int se2_U_inv_checkout(void);
+void se2_U_inv_release(int mem);
+void se2_U_inv_incref(void);
+void se2_U_inv_decref(void);
+casadi_int se2_U_inv_n_in(void);
+casadi_int se2_U_inv_n_out(void);
+casadi_real se2_U_inv_default_in(casadi_int i);
+const char* se2_U_inv_name_in(casadi_int i);
+const char* se2_U_inv_name_out(casadi_int i);
+const casadi_int* se2_U_inv_sparsity_in(casadi_int i);
+const casadi_int* se2_U_inv_sparsity_out(casadi_int i);
+int se2_U_inv_work(casadi_int *sz_arg, casadi_int* sz_res, casadi_int *sz_iw, casadi_int *sz_w);
+int se2_error(const casadi_real** arg, casadi_real** res, casadi_int* iw, casadi_real* w, int mem);
+int se2_error_alloc_mem(void);
+int se2_error_init_mem(int mem);
+void se2_error_free_mem(int mem);
+int se2_error_checkout(void);
+void se2_error_release(int mem);
+void se2_error_incref(void);
+void se2_error_decref(void);
+casadi_int se2_error_n_in(void);
+casadi_int se2_error_n_out(void);
+casadi_real se2_error_default_in(casadi_int i);
+const char* se2_error_name_in(casadi_int i);
+const char* se2_error_name_out(casadi_int i);
+const casadi_int* se2_error_sparsity_in(casadi_int i);
+const casadi_int* se2_error_sparsity_out(casadi_int i);
+int se2_error_work(casadi_int *sz_arg, casadi_int* sz_res, casadi_int *sz_iw, casadi_int *sz_w);
 #ifdef __cplusplus
 } /* extern "C" */
 #endif
