--- apps/rover/src/casadi/se2.c
@@ -1,610 +1,0 @@
-/* This file was automatically generated by CasADi.
-   The CasADi copyright holders make no ownership claim of its contents. */
-#ifdef __cplusplus
-extern "C" {
-#endif
-
-/* How to prefix internal symbols */
-#ifdef CASADI_CODEGEN_PREFIX
-  #define CASADI_NAMESPACE_CONCAT(NS, ID) _CASADI_NAMESPACE_CONCAT(NS, ID)
-  #define _CASADI_NAMESPACE_CONCAT(NS, ID) NS ## ID
-  #define CASADI_PREFIX(ID) CASADI_NAMESPACE_CONCAT(CODEGEN_PREFIX, ID)
-#else
-  #define CASADI_PREFIX(ID) se2_ ## ID
-#endif
-
-#include <math.h>
-
-#ifndef casadi_real
-#define casadi_real double
-#endif
-
-#ifndef casadi_int
-#define casadi_int long long int
-#endif
-
-/* Add prefix to internal symbols */
-#define casadi_f0 CASADI_PREFIX(f0)
-#define casadi_f1 CASADI_PREFIX(f1)
-#define casadi_f2 CASADI_PREFIX(f2)
-#define casadi_s0 CASADI_PREFIX(s0)
-#define casadi_s1 CASADI_PREFIX(s1)
-#define casadi_sq CASADI_PREFIX(sq)
-
-casadi_real casadi_sq(casadi_real x) { return x*x;}
-
-static const casadi_int casadi_s0[7] = {3, 1, 0, 3, 0, 1, 2};
-static const casadi_int casadi_s1[15] = {3, 3, 0, 3, 6, 9, 0, 1, 2, 0, 1, 2, 0, 1, 2};
-
-/* se2_U:(i0[3])->(o0[3x3]) */
-static int casadi_f0(const casadi_real** arg, casadi_real** res, casadi_int* iw, casadi_real* w, int mem) {
-  casadi_real a0, a1, a10, a11, a12, a13, a14, a15, a2, a3, a4, a5, a6, a7, a8, a9;
-  a0=1.0000000000000000e-03;
-  a1=arg[0]? arg[0][2] : 0;
-  a2=fabs(a1);
-  a2=(a0<a2);
-  a3=sin(a1);
-  a4=(a1*a3);
-  a5=2.;
-  a6=cos(a1);
-  a7=1.;
-  a8=(a6-a7);
-  a8=(a5*a8);
-  a4=(a4/a8);
-  a4=(a2?a4:0);
-  a2=(!a2);
-  a8=-1.;
-  a9=casadi_sq(a1);
-  a10=12.;
-  a9=(a9/a10);
-  a9=(a8+a9);
-  a11=casadi_sq(a1);
-  a11=casadi_sq(a11);
-  a12=720.;
-  a11=(a11/a12);
-  a9=(a9+a11);
-  a2=(a2?a9:0);
-  a4=(a4+a2);
-  if (res[0]!=0) res[0][0]=a4;
-  a2=(a1/a5);
-  if (res[0]!=0) res[0][1]=a2;
-  a9=0.;
-  if (res[0]!=0) res[0][2]=a9;
-  a2=(-a2);
-  if (res[0]!=0) res[0][3]=a2;
-  if (res[0]!=0) res[0][4]=a4;
-  if (res[0]!=0) res[0][5]=a9;
-  a9=fabs(a1);
-  a9=(a0<a9);
-  a4=arg[0]? arg[0][0] : 0;
-  a2=(a1*a4);
-  a2=(a2*a3);
-  a11=(a7-a6);
-  a13=arg[0]? arg[0][1] : 0;
-  a14=(a1*a13);
-  a15=(a5*a4);
-  a14=(a14-a15);
-  a11=(a11*a14);
-  a2=(a2+a11);
-  a11=(a5*a1);
-  a14=(a7-a6);
-  a11=(a11*a14);
-  a2=(a2/a11);
-  a2=(a9?a2:0);
-  a9=(!a9);
-  a11=(a13/a5);
-  a14=(a1*a4);
-  a14=(a14/a10);
-  a11=(a11-a14);
-  a14=casadi_sq(a1);
-  a14=(a1*a14);
-  a14=(a14*a4);
-  a14=(a14/a12);
-  a11=(a11+a14);
-  a9=(a9?a11:0);
-  a2=(a2+a9);
-  if (res[0]!=0) res[0][6]=a2;
-  a2=fabs(a1);
-  a0=(a0<a2);
-  a2=(a1*a13);
-  a2=(a2*a3);
-  a3=(a7-a6);
-  a9=(a1*a4);
-  a11=(a5*a13);
-  a9=(a9+a11);
-  a3=(a3*a9);
-  a2=(a2-a3);
-  a3=(a5*a1);
-  a7=(a7-a6);
-  a3=(a3*a7);
-  a2=(a2/a3);
-  a2=(a0?a2:0);
-  a0=(!a0);
-  a4=(a4/a5);
-  a5=(a1*a13);
-  a5=(a5/a10);
-  a4=(a4+a5);
-  a5=casadi_sq(a1);
-  a1=(a1*a5);
-  a1=(a1*a13);
-  a1=(a1/a12);
-  a4=(a4+a1);
-  a4=(-a4);
-  a0=(a0?a4:0);
-  a2=(a2+a0);
-  if (res[0]!=0) res[0][7]=a2;
-  if (res[0]!=0) res[0][8]=a8;
-  return 0;
-}
-
-int se2_U(const casadi_real** arg, casadi_real** res, casadi_int* iw, casadi_real* w, int mem){
-  return casadi_f0(arg, res, iw, w, mem);
-}
-
-int se2_U_alloc_mem(void) {
-  return 0;
-}
-
-int se2_U_init_mem(int mem) {
-  return 0;
-}
-
-void se2_U_free_mem(int mem) {
-}
-
-int se2_U_checkout(void) {
-  return 0;
-}
-
-void se2_U_release(int mem) {
-}
-
-void se2_U_incref(void) {
-}
-
-void se2_U_decref(void) {
-}
-
-casadi_int se2_U_n_in(void) { return 1;}
-
-casadi_int se2_U_n_out(void) { return 1;}
-
-casadi_real se2_U_default_in(casadi_int i){
-  switch (i) {
-    default: return 0;
-  }
-}
-
-const char* se2_U_name_in(casadi_int i){
-  switch (i) {
-    case 0: return "i0";
-    default: return 0;
-  }
-}
-
-const char* se2_U_name_out(casadi_int i){
-  switch (i) {
-    case 0: return "o0";
-    default: return 0;
-  }
-}
-
-const casadi_int* se2_U_sparsity_in(casadi_int i) {
-  switch (i) {
-    case 0: return casadi_s0;
-    default: return 0;
-  }
-}
-
-const casadi_int* se2_U_sparsity_out(casadi_int i) {
-  switch (i) {
-    case 0: return casadi_s1;
-    default: return 0;
-  }
-}
-
-int se2_U_work(casadi_int *sz_arg, casadi_int* sz_res, casadi_int *sz_iw, casadi_int *sz_w) {
-  if (sz_arg) *sz_arg = 1;
-  if (sz_res) *sz_res = 1;
-  if (sz_iw) *sz_iw = 0;
-  if (sz_w) *sz_w = 0;
-  return 0;
-}
-
-/* se2_U_inv:(i0[3])->(o0[3x3]) */
-static int casadi_f1(const casadi_real** arg, casadi_real** res, casadi_int* iw, casadi_real* w, int mem) {
-  casadi_real a0, a1, a10, a11, a12, a13, a14, a15, a16, a17, a18, a2, a3, a4, a5, a6, a7, a8, a9;
-  a0=1.0000000000000000e-03;
-  a1=arg[0]? arg[0][2] : 0;
-  a2=fabs(a1);
-  a2=(a0<a2);
-  a3=sin(a1);
-  a4=(a3/a1);
-  a4=(a2?a4:0);
-  a2=(!a2);
-  a5=1.;
-  a6=casadi_sq(a1);
-  a7=6.;
-  a6=(a6/a7);
-  a6=(a5-a6);
-  a8=casadi_sq(a1);
-  a8=casadi_sq(a8);
-  a9=120.;
-  a8=(a8/a9);
-  a6=(a6+a8);
-  a2=(a2?a6:0);
-  a4=(a4+a2);
-  a2=(-a4);
-  if (res[0]!=0) res[0][0]=a2;
-  a2=fabs(a1);
-  a2=(a0<a2);
-  a6=cos(a1);
-  a8=(a5-a6);
-  a8=(a8/a1);
-  a8=(-a8);
-  a8=(a2?a8:0);
-  a2=(!a2);
-  a10=2.;
-  a11=(a1/a10);
-  a12=casadi_sq(a1);
-  a12=(a1*a12);
-  a13=24.;
-  a12=(a12/a13);
-  a11=(a11-a12);
-  a2=(a2?a11:0);
-  a8=(a8+a2);
-  if (res[0]!=0) res[0][1]=a8;
-  a2=0.;
-  if (res[0]!=0) res[0][2]=a2;
-  a8=(-a8);
-  if (res[0]!=0) res[0][3]=a8;
-  a4=(-a4);
-  if (res[0]!=0) res[0][4]=a4;
-  if (res[0]!=0) res[0][5]=a2;
-  a2=fabs(a1);
-  a2=(a0<a2);
-  a4=arg[0]? arg[0][0] : 0;
-  a8=(a1*a6);
-  a8=(a8-a1);
-  a8=(a8+a3);
-  a11=(a10*a3);
-  a11=(a11*a6);
-  a12=(a11/a10);
-  a8=(a8-a12);
-  a8=(a4*a8);
-  a12=arg[0]? arg[0][1] : 0;
-  a14=(a10*a6);
-  a15=(a10*a6);
-  a15=(a15*a6);
-  a15=(a15-a5);
-  a16=(a15/a10);
-  a14=(a14-a16);
-  a16=1.5000000000000000e+00;
-  a14=(a14-a16);
-  a14=(a12*a14);
-  a8=(a8+a14);
-  a14=casadi_sq(a1);
-  a17=(a5-a6);
-  a14=(a14*a17);
-  a8=(a8/a14);
-  a8=(-a8);
-  a8=(a2?a8:0);
-  a2=(!a2);
-  a14=(a12/a10);
-  a17=(a1*a4);
-  a17=(a17/a7);
-  a14=(a14+a17);
-  a17=casadi_sq(a1);
-  a17=(a17*a12);
-  a17=(a17/a13);
-  a14=(a14-a17);
-  a17=casadi_sq(a1);
-  a17=(a1*a17);
-  a17=(a17*a4);
-  a17=(a17/a9);
-  a14=(a14-a17);
-  a17=casadi_sq(a1);
-  a17=casadi_sq(a17);
-  a17=(a17*a12);
-  a18=720.;
-  a17=(a17/a18);
-  a14=(a14+a17);
-  a2=(a2?a14:0);
-  a8=(a8+a2);
-  a8=(-a8);
-  if (res[0]!=0) res[0][6]=a8;
-  a8=fabs(a1);
-  a0=(a0<a8);
-  a8=-2.;
-  a8=(a8*a6);
-  a15=(a15/a10);
-  a8=(a8+a15);
-  a8=(a8+a16);
-  a8=(a4*a8);
-  a16=(a1*a6);
-  a16=(a16-a1);
-  a16=(a16+a3);
-  a11=(a11/a10);
-  a16=(a16-a11);
-  a16=(a12*a16);
-  a8=(a8+a16);
-  a16=casadi_sq(a1);
-  a5=(a5-a6);
-  a16=(a16*a5);
-  a8=(a8/a16);
-  a8=(-a8);
-  a8=(a0?a8:0);
-  a0=(!a0);
-  a16=(a1*a12);
-  a16=(a16/a7);
-  a10=(a4/a10);
-  a16=(a16-a10);
-  a10=casadi_sq(a1);
-  a10=(a10*a4);
-  a10=(a10/a13);
-  a16=(a16+a10);
-  a10=casadi_sq(a1);
-  a10=(a1*a10);
-  a10=(a10*a12);
-  a10=(a10/a9);
-  a16=(a16-a10);
-  a1=casadi_sq(a1);
-  a1=casadi_sq(a1);
-  a1=(a1*a4);
-  a1=(a1/a18);
-  a16=(a16-a1);
-  a0=(a0?a16:0);
-  a8=(a8+a0);
-  a8=(-a8);
-  if (res[0]!=0) res[0][7]=a8;
-  a8=-1.;
-  if (res[0]!=0) res[0][8]=a8;
-  return 0;
-}
-
-int se2_U_inv(const casadi_real** arg, casadi_real** res, casadi_int* iw, casadi_real* w, int mem){
-  return casadi_f1(arg, res, iw, w, mem);
-}
-
-int se2_U_inv_alloc_mem(void) {
-  return 0;
-}
-
-int se2_U_inv_init_mem(int mem) {
-  return 0;
-}
-
-void se2_U_inv_free_mem(int mem) {
-}
-
-int se2_U_inv_checkout(void) {
-  return 0;
-}
-
-void se2_U_inv_release(int mem) {
-}
-
-void se2_U_inv_incref(void) {
-}
-
-void se2_U_inv_decref(void) {
-}
-
-casadi_int se2_U_inv_n_in(void) { return 1;}
-
-casadi_int se2_U_inv_n_out(void) { return 1;}
-
-casadi_real se2_U_inv_default_in(casadi_int i){
-  switch (i) {
-    default: return 0;
-  }
-}
-
-const char* se2_U_inv_name_in(casadi_int i){
-  switch (i) {
-    case 0: return "i0";
-    default: return 0;
-  }
-}
-
-const char* se2_U_inv_name_out(casadi_int i){
-  switch (i) {
-    case 0: return "o0";
-    default: return 0;
-  }
-}
-
-const casadi_int* se2_U_inv_sparsity_in(casadi_int i) {
-  switch (i) {
-    case 0: return casadi_s0;
-    default: return 0;
-  }
-}
-
-const casadi_int* se2_U_inv_sparsity_out(casadi_int i) {
-  switch (i) {
-    case 0: return casadi_s1;
-    default: return 0;
-  }
-}
-
-int se2_U_inv_work(casadi_int *sz_arg, casadi_int* sz_res, casadi_int *sz_iw, casadi_int *sz_w) {
-  if (sz_arg) *sz_arg = 1;
-  if (sz_res) *sz_res = 1;
-  if (sz_iw) *sz_iw = 0;
-  if (sz_w) *sz_w = 0;
-  return 0;
-}
-
-/* se2_error:(i0[3],i1[3])->(o0[3]) */
-static int casadi_f2(const casadi_real** arg, casadi_real** res, casadi_int* iw, casadi_real* w, int mem) {
-  casadi_real a0, a1, a10, a11, a12, a13, a14, a15, a2, a3, a4, a5, a6, a7, a8, a9;
-  a0=arg[1]? arg[1][2] : 0;
-  a1=arg[0]? arg[0][2] : 0;
-  a0=(a0-a1);
-  a2=sin(a0);
-  a0=cos(a0);
-  a2=atan2(a2,a0);
-  a0=fabs(a2);
-  a3=9.9999999999999995e-08;
-  a0=(a0<a3);
-  a4=1.;
-  a5=-1.6666666666666666e-01;
-  a6=casadi_sq(a2);
-  a5=(a5*a6);
-  a5=(a4+a5);
-  a6=8.3333333333333332e-03;
-  a7=casadi_sq(a2);
-  a7=casadi_sq(a7);
-  a6=(a6*a7);
-  a5=(a5+a6);
-  a5=(a0?a5:0);
-  a0=(!a0);
-  a6=sin(a2);
-  a6=(a6/a2);
-  a0=(a0?a6:0);
-  a5=(a5+a0);
-  a0=casadi_sq(a5);
-  a6=fabs(a2);
-  a6=(a6<a3);
-  a3=5.0000000000000000e-01;
-  a3=(a3*a2);
-  a7=-4.1666666666666664e-02;
-  a8=casadi_sq(a2);
-  a8=(a2*a8);
-  a7=(a7*a8);
-  a3=(a3+a7);
-  a7=1.3888888888888889e-03;
-  a8=casadi_sq(a2);
-  a8=casadi_sq(a8);
-  a8=(a2*a8);
-  a7=(a7*a8);
-  a3=(a3+a7);
-  a3=(a6?a3:0);
-  a6=(!a6);
-  a7=cos(a2);
-  a4=(a4-a7);
-  a4=(a4/a2);
-  a6=(a6?a4:0);
-  a3=(a3+a6);
-  a6=casadi_sq(a3);
-  a0=(a0+a6);
-  a6=(a5/a0);
-  a4=(-a1);
-  a7=cos(a4);
-  a8=arg[1]? arg[1][0] : 0;
-  a9=(a7*a8);
-  a4=sin(a4);
-  a10=arg[1]? arg[1][1] : 0;
-  a11=(a4*a10);
-  a9=(a9-a11);
-  a11=cos(a1);
-  a12=arg[0]? arg[0][0] : 0;
-  a13=(a11*a12);
-  a1=sin(a1);
-  a14=arg[0]? arg[0][1] : 0;
-  a15=(a1*a14);
-  a13=(a13+a15);
-  a9=(a9-a13);
-  a6=(a6*a9);
-  a13=(a3/a0);
-  a4=(a4*a8);
-  a7=(a7*a10);
-  a4=(a4+a7);
-  a1=(a1*a12);
-  a11=(a11*a14);
-  a1=(a1-a11);
-  a4=(a4+a1);
-  a13=(a13*a4);
-  a6=(a6+a13);
-  if (res[0]!=0) res[0][0]=a6;
-  a5=(a5/a0);
-  a5=(a5*a4);
-  a3=(a3/a0);
-  a3=(a3*a9);
-  a5=(a5-a3);
-  if (res[0]!=0) res[0][1]=a5;
-  if (res[0]!=0) res[0][2]=a2;
-  return 0;
-}
-
-int se2_error(const casadi_real** arg, casadi_real** res, casadi_int* iw, casadi_real* w, int mem){
-  return casadi_f2(arg, res, iw, w, mem);
-}
-
-int se2_error_alloc_mem(void) {
-  return 0;
-}
-
-int se2_error_init_mem(int mem) {
-  return 0;
-}
-
-void se2_error_free_mem(int mem) {
-}
-
-int se2_error_checkout(void) {
-  return 0;
-}
-
-void se2_error_release(int mem) {
-}
-
-void se2_error_incref(void) {
-}
-
-void se2_error_decref(void) {
-}
-
-casadi_int se2_error_n_in(void) { return 2;}
-
-casadi_int se2_error_n_out(void) { return 1;}
-
-casadi_real se2_error_default_in(casadi_int i){
-  switch (i) {
-    default: return 0;
-  }
-}
-
-const char* se2_error_name_in(casadi_int i){
-  switch (i) {
-    case 0: return "i0";
-    case 1: return "i1";
-    default: return 0;
-  }
-}
-
-const char* se2_error_name_out(casadi_int i){
-  switch (i) {
-    case 0: return "o0";
-    default: return 0;
-  }
-}
-
-const casadi_int* se2_error_sparsity_in(casadi_int i) {
-  switch (i) {
-    case 0: return casadi_s0;
-    case 1: return casadi_s0;
-    default: return 0;
-  }
-}
-
-const casadi_int* se2_error_sparsity_out(casadi_int i) {
-  switch (i) {
-    case 0: return casadi_s0;
-    default: return 0;
-  }
-}
-
-int se2_error_work(casadi_int *sz_arg, casadi_int* sz_res, casadi_int *sz_iw, casadi_int *sz_w) {
-  if (sz_arg) *sz_arg = 2;
-  if (sz_res) *sz_res = 1;
-  if (sz_iw) *sz_iw = 0;
-  if (sz_w) *sz_w = 0;
-  return 0;
-}
-
-
-#ifdef __cplusplus
-} /* extern "C" */
-#endif
