--- apps/rover/src/casadi/se2.h
@@ -1,65 +1,0 @@
-/* This file was automatically generated by CasADi.
-   The CasADi copyright holders make no ownership claim of its contents. */
-#ifdef __cplusplus
-extern "C" {
-#endif
-
-#ifndef casadi_real
-#define casadi_real double
-#endif
-
-#ifndef casadi_int
-#define casadi_int long long int
-#endif
-
-int se2_U(const casadi_real** arg, casadi_real** res, casadi_int* iw, casadi_real* w, int mem);
-int se2_U_alloc_mem(void);
-int se2_U_init_mem(int mem);
-void se2_U_free_mem(int mem);
-int se2_U_checkout(void);
-void se2_U_release(int mem);
-void se2_U_incref(void);
-void se2_U_decref(void);
-casadi_int se2_U_n_in(void);
-casadi_int se2_U_n_out(void);
-casadi_real se2_U_default_in(casadi_int i);
-const char* se2_U_name_in(casadi_int i);
-const char* se2_U_name_out(casadi_int i);
-const casadi_int* se2_U_sparsity_in(casadi_int i);
-const casadi_int* se2_U_sparsity_out(casadi_int i);
-int se2_U_work(casadi_int *sz_arg, casadi_int* sz_res, casadi_int *sz_iw, casadi_int *sz_w);
-int se2_U_inv(const casadi_real** arg, casadi_real** res, casadi_int* iw, casadi_real* w, int mem);
-int se2_U_inv_alloc_mem(void);
-int se2_U_inv_init_mem(int mem);
-void se2_U_inv_free_mem(int mem);
-int se2_U_inv_checkout(void);
-void se2_U_inv_release(int mem);
-void se2_U_inv_incref(void);
-void se2_U_inv_decref(void);
-casadi_int se2_U_inv_n_in(void);
-casadi_int se2_U_inv_n_out(void);
-casadi_real se2_U_inv_default_in(casadi_int i);
-const char* se2_U_inv_name_in(casadi_int i);
-const char* se2_U_inv_name_out(casadi_int i);
-const casadi_int* se2_U_inv_sparsity_in(casadi_int i);
-const casadi_int* se2_U_inv_sparsity_out(casadi_int i);
-int se2_U_inv_work(casadi_int *sz_arg, casadi_int* sz_res, casadi_int *sz_iw, casadi_int *sz_w);
-int se2_error(const casadi_real** arg, casadi_real** res, casadi_int* iw, casadi_real* w, int mem);
-int se2_error_alloc_mem(void);
-int se2_error_init_mem(int mem);
-void se2_error_free_mem(int mem);
-int se2_error_checkout(void);
-void se2_error_release(int mem);
-void se2_error_incref(void);
-void se2_error_decref(void);
-casadi_int se2_error_n_in(void);
-casadi_int se2_error_n_out(void);
-casadi_real se2_error_default_in(casadi_int i);
-const char* se2_error_name_in(casadi_int i);
-const char* se2_error_name_out(casadi_int i);
-const casadi_int* se2_error_sparsity_in(casadi_int i);
-const casadi_int* se2_error_sparsity_out(casadi_int i);
-int se2_error_work(casadi_int *sz_arg, casadi_int* sz_res, casadi_int *sz_iw, casadi_int *sz_w);
-#ifdef __cplusplus
-} /* extern "C" */
-#endif
