+++ apps/rover/src/parameters.h
@@ -1,0 +1,30 @@
+#ifndef E00B391E_08D8_47AC_B779_57B422AF3850
+#define E00B391E_08D8_47AC_B779_57B422AF3850
+
+// gains
+//const double gain_heading = 0.5;
+//const double gain_cross_track = 0.4;
+//const double gain_along_track = 1.0;
+
+// parameters
+//const double wheel_radius = 0.0365;
+//const double wheel_base = 0.2255;
+//const double wheel_separation = 0.4;
+
+//#define STEERING_ACKERMANN
+
+// gains
+const double gain_heading = 10;
+const double gain_cross_track = 10;
+const double gain_along_track = 10;
+
+// parameters
+const double wheel_radius = 0.108;
+const double wheel_base = 0.667;
+const double wheel_separation = 0.667;
+
+#define STEERING_DIFFERENTIAL
+
+#endif /* E00B391E_08D8_47AC_B779_57B422AF3850 */
+
+
